pyproject.toml
[tool.poetry]
name = "syn-pg"
version = "0.1.0"
description = "A Python application for generating synthetic data using GPT and PostgreSQL."
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.10"
psycopg2 = "^2.9.6"
python-dotenv = "^1.0.0"
openai = "^0.27.4"
apscheduler = "^3.9.1"

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"
pytest-mock = "^3.10.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core>=1.1.0"]
build-backend = "poetry.core.masonry.api"